# Conflicts:
#	index.js
#	src/pages/all-brands/allBrands.js

+++ src/pages/color/colorTypes/colorTypeFunctions.js
@@ -2,14 +2,36 @@ var URL = "http://localhost:8080/api/color-types"
 let router;
 
 export function initColorTypes(navigoRouter) {
-
-    document.getElementById("submit").onclick = addColorType;
     getAllColorTypes();
     router = navigoRouter
+    document.getElementById("submit").onclick = addColorType;
     document.getElementById("table").onclick = (element) =>{
         let id = element.target.id
-        deleteColorType(id);
+        if(id.includes("delete")) {
+            deleteColorType(id);
+        }
     }
+    document.getElementById("table").ondblclick = (element) =>{
+        let id = element.target.id
+        if(id.includes("text")){
+            changeTdToInput(id);
+        }
+        var el = document.getElementById(id);
+        el.addEventListener("keydown", function(event) {
+            if (event.key === "Enter") {
+                console.log("Enter clicked "  + id)
+                let field = document.getElementById(id)
+                field.readOnly = true
+                editColorType(id)
+            }
+        });
+    }
+
+}
+
+function changeTdToInput(id){
+    const field = document.getElementById(id)
+    field.readOnly = false
 }
 
 
@@ -18,13 +40,13 @@ async function getAllColorTypes() {
     document.getElementById("tbody-all").innerHTML = ""
     try{
         const data = await fetch(URL).then(res => res.json());
-        console.log(data)
+        //<td id="color-type-id-${colorType.id}" value="${colorType.id}">${colorType.id}</td>
         const tableRowsArray = data.map(
             (colorType) =>
                 `
         <tr>
-            <td>${colorType.id}</td>
-            <td>${colorType.type}</td>
+            
+            <td><input readonly type='text' id="text${colorType.id}" value='${colorType.type}'></td>
             <td><button id="delete${colorType.id}">Delete</button></td>
         `
         );
@@ -38,7 +60,6 @@ async function getAllColorTypes() {
 
 async function addColorType() {
     const type = document.getElementById("if1").value;
-    console.log(type)
     const newColorType = {
         type
     };
@@ -55,50 +76,41 @@ async function addColorType() {
 
 }
 
-
-
-
 async function deleteColorType(idToDelete) {
-    console.log(idToDelete)
-    if(idToDelete.includes("delete")){
-        idToDelete = idToDelete.split('delete')[1]
+    idToDelete = idToDelete.split('delete')[1]
+    var r = confirm("If you delete this Color Type all associated color mixes will be deleted.");
+    if (r==true)
+    {
         const response = await fetch(URL + "/" + idToDelete, {
             method: "DELETE",
 
         }).then((res) => res.json())
-        console.log(response)
         location.reload();
     }
-
-
 }
 
-
-
-/*
-async function editColorMix() {
-    const colorMixId = document.getElementById("if4").value;
-    const colorCode = document.getElementById("if1").value;
-    const colorName = document.getElementById("if2").value;
-    const colorTypeId = document.getElementById("if3").value;
-
-    const editedColorMix = {
-        colorMixId,
-        colorCode,
-        colorTypeId,
-        colorName
+async function editColorType(idFromJs) {
+    const id = idFromJs.split('text')[1]
+    const type = document.getElementById(idFromJs).value;
+    const editedColorType = {
+        id,
+        type
     };
-
-    console.log(editedColorMix)
-    const id = await fetch(URL, {
+    const data = await fetch(URL, {
         method: "PUT",
         headers: {
             "Content-Type": "application/json",
         },
-        body: JSON.stringify(editedColorMix),
+        body: JSON.stringify(editedColorType),
     })
         .then((res) => res.json())
+    if(document.getElementById(idFromJs).value === data.type){
+        document.getElementById(idFromJs).style.boxShadow = "0px 0px 20px 1px #00FF00";
+    }else {
+        document.getElementById(idFromJs).style.boxShadow = "0px 0px 20px 1px #FF0000";
+    }
+    setTimeout( () =>{
+        document.getElementById(idFromJs).style.boxShadow = "none";
+    }, 2000);
+
 }
-*/
-
-
